+++ core/include/update.inc.php
@@ -829,28 +829,20 @@ if ($this->getData(['core', 'dataVersion']) < 11400) {
 	// Redirection des pages d'administration vers la bannière de connexion
 	$this->setData(['config', 'connect', 'redirectLogin', true]);
 
-	// Transforme les URL en références relatives
-	/*
-	$baseUrl = $this->getData(['core', 'baseUrl']);
-	$baseUrl2 = str_replace('?', '', $baseUrl);
-	foreach ($this->getHierarchy(null,null,null) as $parentKey=>$parentValue) {
-		$pageList [] = $parentKey;
-		foreach ($parentValue as $childKey) {
-			$pageList [] = $childKey;
-		}
-	}
-	foreach ($pageList as $parentKey => $parent) {
-		$s = $this->getPage(  $parent, self::$i18n);
-		// Suppression des sous-dossiers
-		$s = str_replace ($baseUrl, './', $s);
-		$s = str_replace ($baseUrl2, './', $s);
-		$this->setPage( $parent, $s, self::$i18n);
-	}
-	*/
+
 
 	// Suppression de la variable URL dans core
 	$this->deleteData(['core', 'baseUrl']);
 
 	// Mise à jour
 	$this->setData(['core', 'dataVersion', 11400]);
+}
+
+// Version 12.0.00
+if ($this->getData(['core', 'dataVersion']) < 12000) {
+	// Supprime un cookie non nécessaire
+	helper::deleteCookie('ZWII_USER_LONGTIME');
+
+	// Mise à jour
+	$this->setData(['core', 'dataVersion', 12000]);
 }