+++ CHANGES.md
@@ -12,6 +12,7 @@
 - Mise à jour automatisée, affichage de l'erreur en cas d'échec.
 - Suppression du support de l'import à partir d'une version 9, y compris pour la restauration des sauvegardes.
 - Optimisation du chargement des base de données, mise à jour des scripts jsonDB et dot.
+- Connexion persistante, l'activation de la case à cocher *Rester connecté sur ce navigateur* ne ferme pas la session lorsque le navigateur est fermé. 
 
 ## Version 11.4.02
 ### Modifications :